zeedit: add new options from `print_items` to the example configuration
The configuration file is used by the `zeedit.py` script
and can be modified as required by the user.

The new variables are `sm_blocks`, `sm_size`, and `sm_sanitize`,
which are passed as the corresponding arguments `blocks`, `size`,
and `sanitize` of `printf.print_summary`.

The other variables are re-ordered to follow the same order
as they defined in `printf.print_summary`.