print: make sure the fields printed don't have a trailing separator
Depending on the arguments some fields will be omited when printing
information with `print.print_items`.
If they are omitted no final separator (`;`) should appear at the end
of the field.

Therefore, we print the seaparator at the beginning of the field
instead of at the end, so if the field is omitted there is no separator
either.