search: new `repost` argument to avoid resolving reposts
With `repost=True` the claim returned by `search.search_item`
will be the original claim if the claim is a repost.

With `repost=False` the claim being searched
will not be further resolved so it will be returned
just as is, even if it's a repost.

If it's a repost it won't be able to be downloaded
by `download.download_single` (which calls `lbrynet get`)
as reposts cannot be downloaded by themselves.