download: new `repost` argument to avoid resolving reposts
With `repost=True` the claim returned by `search.search_item`
will be the original claim if the claim is a repost.
Then it can be downloaded by `download.download_single`.

With `repost=False` the claim being searched
will not be further resolved so it will be returned
just as is, even if it's a repost.

The repost won't be downloaded as a repost can't be downloaded.
Some information will be printed about it but it will be skipped.