download_multi: new `repost` argument to avoid resolving reposts
With `repost=True` the claim returned by `search.search_item`
will be the original claim if the claim is a repost.
Then it can be downloaded by `download.download_single`,
which is used by `download_multi.ch_download_latest`,
and in turn by `download_multi.ch_download_latest_multi`.

With `repost=False` the claims being searched
will not be further resolved so they will be returned
just as is, even if they are reposts.

Reposts won't be downloaded as a repost can't be downloaded.
Some information will be printed about them but they will be skipped.