sort: new `reverse` parameter to show newer items first
By default the `sort.sort_items` and `sort.sort_invalid`
methods order the elements of the output lists in ascending order,
meaning that the oldest claims by `release_time` come first,
and the newest items appear at the end.

With the new option `reverse=True` the newest claims
will be first.

This option in turn is passed to `print.print_items`
and `printf.print_summary` to control how the claims
are printed to the terminal or to a local file.