sort: new method `sort_items_size` to calculate the size of the files
In the new method `sort.sort_items_size` the output
of `sort.sort_items` or `sort.sort_invalid` is passed to
a modified `search_utils.downloadable_size` so that it calculates
the total size of all claims if all their blobs are downloaded.

The method will return a dictionary with all found claims
and their total size.

The `printf.print_summary` method is updated to use this new method
so that it displays the total size at the same time that it displays
the number of claims.

The new method `sort_items_size` and the existing method `print_items`
are made available in the toplevel `lbrytools` namespace
so that they can be called by external applications
that use this library.