search_utils: add `duration` to return dictionary of `downloadable_size`
The `search_utils.downloadable_size` method calculates the size in bytes
of the input list, but now it also calculates the duration in seconds.

This is for those claims that have a duration, like video and audio streams.
For other claims, the duration is just zero, so nothing is added
to the final duration.

The return value of `downloadable_size` is now a dictionary with two
keys `'size'` and `'duration'`, instead of a single integer for the size.

This means that the methods which use this function also have to be updated.

One method is `search_utils.sort_filter_size`, which now displays
the duration in hours and days.
Now it returns as output value a dictionary with three keys,
`'claims'`, `'size'`, and `'duration'`.

The other method is `sort.sort_items_size` which also returns
a dictionary with three keys, `'claims'`, `'size'`, and `'duration'`.