blobs_ratio: move `get_data_dir` to `funcs.py` module
Instead of defining `get_data_dir` inside `blobs_ratio.py`
it is placed in `funcs.py` so that it can be called
by other methods.

The `server_exists` method is also removed from `blobs_ratio.py`,
as it already exists in `funcs.py`.

The final printing of the blobs ratio information is now printed
using the `funcs.print_content` method.

Also change the parameter `plot_hst` to `False` so that the histogram
is not created by default.