support: add `invalid` arguments to `list_supports`
If the claim has been removed from the blockchain (output spent),
then it won't be able to be resolved. However, a support may still exist
for that invalid claim (channel or stream).

If `invalid=True` it will only print these invalid claims.

If `invalid=False` it will print all claims including the invalid
ones. The invalid claims will appear with brackets around the name
of the claim, like "[@InvalidChannel]".

In this case the amount of support in the claim is only our support.