support: use the new method `get_all_supports`
The new method provides all supports in a dictionary.
It can provide all supports, or the supports for only valid
or only invalid claims.

For the `list_supports` method, we use all supports,
so that we can show also the invalid claims in certain way
depending on the value of the resolved value.