support: new `abandon_support_inv` to abandon supports on invalid claims
The method accepts an `all_supports` dictionary that has information
on all supports, both valid and invalid.

If the given claim ID (`cid`) or `name` passed to the function
exists in one of the lists of supports for invalid claims,
we can handle it, without needing to resolve it online.

The support can be removed completely, or it can be changed
to another amount (`keep`); given that it is an invalid claim
the support should probably be removed completely.