support: new `calculate_abandon` to abstract abandoning supports
The `support_abandon` method is encapsulated in `calculate_abandon`
in order to store the results of the command into a dictionary,
and also prepare the output text.

This method will be used to display the results of abandoning
or changing the support of regular claims and also invalid claims
(those that cannot be resolved online because the output was spent).