zeedit: new repository separate from `lbrytools`
The `zeedit` program uses the `lbrytools` library.
Originally, `zeedit` was hosted in the same `lbrytools` repository,
but since it is not really part of this library,
starting with this commit it is removed from this repository,
and moved into its own repository.

All previous commits before this are shared with the new `zeedit` repository
so that it is possible to inspect the development history of `zeedit`
alongside `lbrytools`.

Now `zeedit` can be installed by itself, and `lbrytools`
can be installed as a dependency, for example, as a Git submodule,
or as an external library.

Therefore, the `zeedit` code is removed completely
from this repository, and only the generic library code remains.

The `README.md` is updated to reflect the changes.