peers: new method to list the peers that are in the network
The method `peers.list_peers` returns the peers
that exist in the LBRY network for the latest claims
of a given channel.

The method uses threads to search multiple claims in parallel
which significantly reduces the peer search time.

By default it uses 32 threads, which can be increased or decreased
depending on CPU power.

Also update `README.md`, `__init__.py`,
and the documentation `lbrytools.md`.