funcs: new method to process a list of channels and numbers
A channel name can be accompanied by a number, which represents
something like the number of claims belonging to that channel.

This pair can be put inside of a list with other pairs
so that another function goes through the list, operating on the channels.

This can be used to download or delete many claims from many channels
at the same time.
```
channels = ["Channel",
            ["@Ch1"],
            ["MyChan", 2]]
```

The new method `funcs.process_ch_num` makes sure that the list
is correctly formed, with each element having a pair of values.
The channel name must be a string, and the number must exist,
or should have a default value.
```
channels = [{"channel": "@Channel", "number": 2},
            {"channel": "@Ch1", "number": 2},
            {"channel": "@MyChan", "number": 2}]
```