peers_multi: new method to list the peers of multiple channels
The method `peers_multi.list_ch_peers` takes as input
a list of channels with the number of claims
that will be searched for peers, and then it calls `peers.list_peers`
for each channel.

The method uses threads to search multiple channels in parallel
which significantly reduces the peer search time.

By default it will search channels using 8 threads;
for `peers.list_peers` it will use 32 threads
to search for peers in the claims.

Also update `README.md`, `__init__.py`,
and the documentation `lbrytools.md`.