channels: new function `list_ch_subs` to list subscriptions
The method `channels.channel_subs` is removed.
In its place the new method `channels.list_ch_subs`
will perform the same task of printing the list of subscribed channels.

Also update `README.md`, `__init__.py`,
and the documentation `lbrytools.md`.

----

The new method depends on new methods, `channels.search_ch_subs`
which performs the actual search, and `channels.print_ch_subs`
which prints the result of the previous method.

The new method `list_ch_subs` has the parameter `show_all`
to show all subscribed channels listed in the wallet,
regardless of whether they actually exist or not.

The parameter `filtering` can be used to control how to filter
the list of channels: 'valid', 'notifications', 'both'.
This only works when `show_all=False`.

The parameter `valid` filters the list of channels by those
that are valid and resolve online (True), or by those that don't
exist or are invalid (False).
This only works when `show_all=False` and `filtering` is `'valid'`
or `'both'`.

The parameter `notifications` filters the list of channels by those
that have notifications (True), or by those that don't (False).
This only works when `show_all=False` and `filtering` is `'notifications'`
or `'both'`.

If `show_all=False` and `filtering='both'` it will first filter
the channel by those that are valid, and then by those with notifications.