peers_multi: new method to list peers from subscribed channels
The method `peers_multi.list_ch_subs_peers` first uses
`channels.list_channel_subs` to get the list of channels
to which we are subscribed; then `peers_multi.list_ch_peers`
is called on the full or partial list of channels.

Internally each channel will be passed to `peers.list_peers`
to find the peers for the claims in the channel.

By default it will search channels using 32 threads;
for `peers.list_peers` it will use 16 threads
to search for peers in the claims.

Also update `README.md`, `__init__.py`,
and the documentation `lbrytools.md`.