download_multi: use the new `process_ch_num` method to parse channels
In the `ch_downloaded_latest_multi` method we parse the list of channels
by using the new method `funcs.process_ch_num`.

This will take an input like this
```
channels = ["Channel",
            ["@Ch1"],
            ["MyChan", 2]]
```

and produce a proper output like this
```
channels = [{"channel": "@Channel", "number": 2},
            {"channel": "@Ch1", "number": 2},
            {"channel": "@MyChan", "number": 2}]
```

This is then unpacked in a loop and the two arguments,
`channel` and `number`, can be passed to the actual download method
`ch_download_latest`.

----

The `print.print_multi_list` will be updated as well to use
the `funcs.print_content` method.