search_ch: allow passing a channel within `[brackets]`
Some channels may be passed with brackets surrounding them,
like `[@channel]`. This typically indicates that the channel
is invalid, and no longer resolves online.

If a channel like this is input, We will remove the brackets
and pass the internal name, `@channel`.

The search will probably fail as we expect from the brackets
that the channel is invalid.