accounts: improved structure of the code
The code of `accounts.list_accounts` is now divided into various
subfunctions. First `get_wallet_info` gets the information on
the wallet and its accounts, and `print_accounts` prints
the actual information.

The function `get_wallet_info` calls `get_bal_wallet` to get
the summary balance of the total wallet; then it calls `get_accounts`
to get the balance of each account in the wallet.

The `get_bal_wallet` function gets the first wallet from `get_wallets`,
which is normally `'default_wallet'`, and from it it extracts
the total amount of credits, available credits, and reserved credits;
as part of the reserved credits, it gets those used for claims,
for supports, and coming from tips.

The `get_accounts` function gets all accounts in the loaded wallet,
then collects the 'id', 'name', 'generator', as well as the summary
of credits of each account. It also searches all the sub-addresses
of each account to see the addresses, and the how many times
they have been used.

The `accounts.list_accounts` method now accepts the `wallet_id`
parameter which allows us to specify the wallet to load.
By default it loads the wallet with ID 'default_wallet'
which is normally created by `lbrynet` when run the first time.

The documentation `lbrytools.md` is updated as well.