comments: new `update_comment` to modify comments
It takes a `comment_id` as input, then it gets the channel
associated with that ID, then signs the new comment
with that channel's key if it actually exists in our wallet.

If the channel corresponding to that comment doesn't belong to us
then we won't be able to sign it, and we won't be able to edit
the comment.

Finally, print the information on the comment that was successfully
edited.

----

Add the method to the `__init__` file so that the method is available
from the toplevel `lbrytools` namespace.
Also update the documentation in `README.md` and `lbrytools.md`.