comments: use `sign_comment` to create the actual comment
In order to post a comment on Odysee servers, the comment
needs to be signed with the private key of channel,
which is contained in the wallet that is read by `lbrynet`.

Now we use the new method `sign_comment` to sign the message
before actually sending the transaction to the comment server.
With this signature, the comment should be correctly created.

Then we get the response of the comment server and print information
about it, like the timestamp of the comment, and properties
such as support, currency, fiat, hidden, pinned, etc.

----

Also the `wallet_id` is provided as a new parameter because
to sign the data, we need to know the wallet that holds
the channel keys.

----

Also, we add the default value of `"(None)"` to the title
in case the claim has no title.

----

Add the method to the `__init__` file so that the method is available
from the toplevel `lbrytools` namespace.
Also update the documentation in `README.md` and `lbrytools.md`.