comments: new `abandon_comment` to remove comments we have created
It takes a `comment_id` as input, then it gets the channel
associated with that ID, then signs the `comment_id`
with that channel's key if it actually exists in our wallet.

If the channel corresponding to that comment doesn't belong to us
then we won't be able to sign the `comment_id`, and we won't be able
to delete the comment.

Finally, print the information on the comment that was sucessfully
abandoned.

----

Add the method to the `__init__` file so that the method is available
from the toplevel `lbrytools` namespace.
Also update the documentation in `README.md` and `lbrytools.md`.