comments: new `sign_comment` to sign comments with a channel's key
In order to post a comment on the comment server,
the comment needs to be signed with the private key of the channel.
The key is contained in the wallet file that is loaded
by `lbrynet`.

The message is encoded in binary, and then it is turned
into a hexadecimal value:
```
hexdata = data.encode().hex()
```

Then the hexidecimal data is signed by `channel_sign`:
```
lbrynet channel sign --channel_name=@some-channel --hexdata=hexdata
```