comments: new `print_cmnt_result` to print the comment answer
This method will be used when the comment is created or edited.

It prints information returned from the comment server
including timestamp, signing timestamp, comment ID, parent comment ID,
currency, value of support, variables of whether the value is fiat,
the comment is hidden, or pinned.

And it also prints the comment itself.

----

Also set `fdate=False` by default.