comments: new `sanitize` parameter for `list_comments`
By default, the comments are shown with all emojis intact,
however, if we want to use the output in programs that don't suppor emojis
we can remove them with the `sanitize` option.

This requires the `emoji` Python package being installed.

----

Also the `funcs.sanitize_name` method is exposed in the toplevel
`lbrytools` namespace so it can be used by other functions.