comments: fix small bugs in the comment code
For `list_comments`, comments that are empty would cause
an error in `print_r_comments` as the code assumes
the length of the string to be at least 1.

Now it can handle completely empty strings with length 0.
This is a rare case as normally comments must have
at least one character in order to be created.

----

For `update_comment` return in case of empty `comment` or `comment_id`.

For `abandon_comment` return in case of empty `comment_id`.