publishes: store the `canonical_url` as `name`
In the `get_channel_claims` method we store the name of the channel
when getting the claims of one of that channels that we own.

Originally we store the `'name'` but this doesn't include
any part of the claim ID, for example, `@channel`.

Now we store the `'canonical_url'` without the intial `lbry://` part,
so that at least we have part of the claim ID, and we can differentiate
two channels that have the same name, for example, `@channel:5`
and `@channel:f`.

In `print_channel` we print the same value, the `'canonical_url'`
without the initial `lbry://` part.