config, status: new methods to show the status and settings of `lbrynet`
The new methods are `list_lbrynet_status` and `list_lbrynet_settings`
which are basically wrappers aroung the `lbrynet status` and
`lbrynet settings_get` methods.

Some of the returned fields are formatted depending on whether the data
is a string, an integer, a float, or others.
The information is printed to the terminal or to a file.

For `list_lbrynet_settings` each key is mostly printed individually.

For `list_lbrynet_status` various keys are grouped together
in paragraphs and printed together.