blobs: small changes in `count_blobs_all` function
Now the `blobfiles` directory is by default the value
returned by `funcs.get_bdir`, which is taken from `settings_get`.

New `insubfunc=True` input parameter is passed to `count_blobs`
in order to omit displaying a message when processing each claim.

Small changes in the name of the variables and the spacing of the code.