blobs: small changes in `count_blobs` function
Now the `blobfiles` directory is by default the value
returned by `funcs.get_bdir`, which is taken from `settings_get`.

New `insubfunc` input parameter in order to control whether
to display a message or not.

Now we use the `with` statement to open and close the `sd_hash` file
with the list of blobs.

Small changes to the printing of the information on the blobs.