blobs: in `count_blobs_all` use new `c_blobs` method
Instead of using `count_blobs` to count the blobs
and print the result, we use the internal functions `c_blobs`
to only count the blobs, and then `prnt_blobs` to print the result,
if `print_msg=True`.

Also add spacing if there is an error in counting, that is,
if a claim is not found (maybe invalid claim),
or when the claim was downloaded but the `sd_hash` blob
is not found in the `blobfiles` directory (requires a redownload).