blobs: new method `prnt_blobs` to print the information
The information is searched by the `count_blobs` method
but only at the end it is printed by `prnt_blobs`
which unpacks the input dictionary.

The optional parameter `print_each` determines whether
to print each individual blob hash or just a summary
for a given claim.

If the claim is missing the `sd_hash` file, it won't be possible
to open it, so the number of blobs won't be counted and displayed.