support: display full name of the claim if it can be resolved
The base name of a single claim or channel can be repeated,
meaning that `@channel` could be interpreted to be
`@channel#a`, `@channel#3`, `@channel#b7e` or another channel.

To avoid confusion, we need to show this full name
when displaying the corresponding support.

The additional characters after the `#` symbol correspond to the claim ID.
For claims that are invalid and no longer resolve online, the `claim_id`
no longer exists, therefore we can only show the base name `@channel`.