support: change the calculation of the `existing_support`
Instead of re-assingning the `resolved` variable that has the resolved
information of the current claim, we use other variables
to make the calculation more obvious.

Now we make it clear that `existing_support` is the sum
of a base support and an additional support:
```
existing_support = base + supp
```

For claims that are invalid and cannot be resolved online
there is no `base` support anymore; any additional support `supp`,
if it exists, may come from us.

We should remove this support given that it no longer boosts
any valid claim.

Also, in case of invalid claims, specify an empty `meta` dictionary
instead of taking it from `resolved["meta"]`.