support: show `'title'` of the claim if it exists
If the claim can be resolved online, we can inspect the `'value'`
and `'title'` keys of the claim, and add the title to the output.

If the claim cannnot be resolved probably it has been removed,
so neither `'value'` nor `'title'` will be found.

In this case, the `title` is just the same `name`
but surrounded by brackets to indicate it is invalid.