support: add the `resolved` dictionary inside the `support` dictionary
The `get_all_supports` method returns a big dictionary
with all the support information.

Previously the information was split into support and resolved:
- 'all_supports': list with the `support` dictionaries only
- 'all_resolved': list with the `resolved` dictionaries only

Now the `resolved` dictionary is included inside each `support` dictionary,
under the `'resolved'` key, so the `all_resolved` list is no longer needed,
and is removed.

For invalid claims, the `'resolved'` key will have a value of `False`.

----

The same is done for valid supports:
- 'valid_supports': list with only `support` dictionaries for valid claims
- 'valid_resolved': list with only `resolved` dictionaries for valid claims

The second of these lists is removed, as the resolved information
is included in the `'resolved'` key.

---

The last list remains, 'invalid_supports'.
In this case, the `'resolved'` key will always be `False` for each element.