support: place the output of `calculate_abandon` in a single dictionary
Previously the output was a tuple with two items
- calc: a dictionary
- text: a list of strings

Now it's a single dictionary which includes the same `calc` dictionary
as previously, and the `text` list is added as the `'text'` key.

The two functions `abandon_support` and `abandon_support_inv`
are updated to accept the new output of `calculate_abandon`.