support: use internal wrapper `search_cid_th` instead of `search_item`
Inside `get_all_supports` the internal wrapper `search_cid_th`
is used for both the threaded search and the non-threaded search.
The wrapper leaves an empty line if a given claim is not found (invalid claim)
just to have a better output in the terminal.

Otherwise the function works the same.