search_utils: expand output dictionary of `downloadable_size`
Given as input a list of claims, `downloadable_size`
would only return the total `size` in bytes,
and the `total_duration` in seconds of all claims.

Normally these values are converted into gibibytes,
and into hours, minutes, seconds, and days by the calling functions.
So instead of doing these calculations elsewhere
we do it in this method already.

Besides the existing two keys:
- `'size'`: total bytes of the list of claims
- `'duration'`: total seconds of the list of claims

We extend the output dictionary with six more keys:
- `'size_GB'`: total size in GiB (floating point value)
- `'d_h'`: integer hours HH when the duration is shown as HH:MM:SS
- `'d_min'`: integer minutes MM when the duration is shown as HH:MM:SS
- `'d_s'`: integer seconds SS when the duration is shown as HH:MM:SS
- `'days'`: total seconds converted into days (floating point value)
- `'text'`: text describing the number of claims, the total size in GiB,
  and the total duration expressed as HH:MM:SS, and days