printf: add `threads` parameter to `print_summary`
Since the internal method `sort.sort_items_size` now can use
threads to search various claims in parallel, the `threads`
argument is added to `print_summary`, which is then passed
to `sort.sort_items_size` and `sort.sort_invalid`.

The `threads` parameter is only used with `invalid=True`
to list invalid claims, as to determine if a claim is invalid
it needs to be resolved online, which is a relatively slow process.

When `invalid=False` the claims are not resolved,
and the output is returned very fast from `sort.sort_items`,
which calls `lbrynet file list` to return the list of claims
from the local database only.

----

Also update the `lbrytools.md` documentation.