printf: count time it takes to search items before printing
When searching claims with `sort_items_size`
it may take some time to return, especially if `invalid=True`.
Therefore we output the time to more or less see how long it takes
to get the claims.

The `sort_items_size` method doesn't need to print anything itself.
The printing can be done by the calling functions,
such as `print_summary`.