s_wallet: new method `sync_wallet` to download and synchronize online wallet
If a LBRY wallet is created in an online service such as Odysee,
we can download that wallet, and synchronize it with our local
LBRY wallet which was set up by `lbrynet`.

To do this we need to do various intermediate steps.

1. First we need to get one `auth_token` from Odysee servers.
2. Then we need to validate the token with the email and password
   of our account on Odysee.
3. Then we get a `sync_hash` from the local wallet.
4. Then we use the validated `auth_token` and the `sync_hash`
   to get the online wallet `sync_data`.
5. Then we pass the wallet sync data to the `lbrynet sync apply`
   method to perform the synchronization.

The 5th step is optional, only if the `sync=True` parameter
is provided to `sync_wallet`.

We can provide optional parameters to determine the wallet to synchronize,
if we have multiple wallets, and also the online address of the servers
for Odysee.

We return a big dictionary with all intermediate information
- `'auth_token'`: token
- `'validated'`: validated online account
- `'sync_hash'`: sync hash of the wallet
- `'sync_data'`: input sync data for the wallet
- `'sync_result'`: output sync data of the wallet