print_ch: use `funcs.print_content` to print the simple format
With `print_channels`, when `simple=True` it will join all channel strings
into a single long string, and print that.

The string is put inside a list (a single element list), and then
we print this list using `funcs.print_content`.

The `funcs.print_content` method already opens the file descriptor
and closes it when finished so we don't need to do it manually,
and thus this code is moved above the opening of the file descriptor.