print_ch: use `funcs.print_content` to print the three column format
With `print_channels`, when `simple=False` it will print the channel
strings arranged in three columns, meaning that each row will have
three elements at most.
```
  c1   c2   c3
  c4   c5
```

Instead of printing each row immediately we put each row
inside a `line` variable, and then we append this variable to the `out` list.
Finally we call `funcs.print_content` to print the list.

Since `funcs.print_content` already opens the file descriptor and closes it
when finished, we remove the portion of the code that does this manually.