print_ch: move code around to avoid repeating the time printing
With `print_channels`, no matter if `simple` is `True` or `False`,
the final time that the method ran is always printed
after showing the found channels.

So printing of the time is removed form the `if simple` test,
and left at the end of the method.

The test is simplified, `if simple` print the single long string;
else print the three column format.