blobs: add a new empty line if the item was not found in the blob search
With default arguments `count_blobs_all` goes through all claims
and searches blobs of all of them.

For invalid items, it should print an empty line if the internal
function `c_blobs` is used a subfunction, which is the case with
`count_blobs_all`.

This empty line just makes the information displayed to the terminal
easier to read.

----

Also update the docstring for the `print_msg` and `print_each`
parameters of `count_blobs_all`. The first parameter, `print_msg`,
defaults to `False`; previously it incorrectly stated
that its default value was `True`.