peers: update the docstring of `search_ch_peers` and `list_peers`
The dictionary output of `calculate_peers` was augmented
with two new keys, `'peers_tracker'` and `'peers_user'`.

This output is added to the `'streams_info'` key
of the `base_peers_info` variable, which is then processed
and augmented by `process_ch_peers`.

The output of `process_ch_peers` is then the output
of `search_ch_peers` and `list_peers`.