peers: rename `list_peers` to `list_ch_peers`
Since this method specifically searches for the claims
of a specific channel, and then searches the peers in those claims,
we make sure that this method includes `ch` in the name.

The already existing method `peers_multi.list_ch_peers`
needs to be renamed `peers_multi.list_chs_peers` as it accepts
multiple channels and not only one.

----

Rename the corresponding methods in the documentation `README.md`
and `lbrytools.md`.