peers_multi: add method `process_chs_peers` to process the peer search
Previously the output of `ch_search_ch_peers` would be a list of
dictionaries returned by calling `peers.search_ch_peers`
on each of the channels passed to the `ch_search_ch_peers` method.

This output would then be passed directly to the printing methods
`print_ch_p_lines` and `print_ch_p_summary`.

However, since `peers.search_ch_peers` now distinguishes
between user peers and tracker peers, now we have more information
that can be presented in the output.

Therefore we add a new method `process_chs_peers` to manipulate
the results of the raw peer search for all channels, and return
a more complete dictionary to be used with the printing functions.

The keys in the new output dictionary are:
- `'base_chs_peers_info'`: original output, list of dictionaries returned
   by multiple `peers.search_ch_peers`
- `'n_channels'`: total number of channels
- `'chs_n_streams'`: total number of downloadable streams in all channels
- `'chs_total_size'`: total size of all streams searched in all channels
- `'chs_total_duration'`: total duration of all streams searched in all channels
- `'chs_streams_with_hosts'`: total number of streams with at least one user host in all channels
- `'chs_streams_with_hosts_all'`: total number of streams with at least one host in all channels
- `'chs_total_peers'`: total number of user peers in all searched streams in all channels
- `'chs_total_peers_all'`: total number of peers in all searched streams in all channels
- `'chs_unique_nodes'`: list of unique user peers in all searched streams in all channels
- `'chs_unique_trackers'`: list of unique tracker peers in all searched streams in all channels
- `'chs_peer_ratio'`: average number of user peers per stream per channel
- `'chs_peer_ratio_all'`: average number of total peers per stream per channel
- `'chs_hosting_coverage'`: average user hosting coverage per channel
- `'chs_hosting_coverage_all'`: average hosting coverage per channel
- `'chs_local_node'`: boolean value that indicates if we are hosting at least one stream

----

Since now `ch_search_ch_peers` returns a new output dictionary,
the consuming methods `print_ch_p_lines` and `print_ch_p_summary`
need to be slightly modified to extract the required quantities
out of the `'base_chs_peers_info'` key, which has the original
expected output (a list of dictionaries).

----

The import is changed from `lbrytools.peers as peers`
to `lbrytools.peers as prs` just to avoid using the word `peers`
as a module name.