peers_ch: move `list_ch_peers` to new module
Since this method searches for the claims of a specific channel,
and then searches for the peers in those claims, we remove
this method from `peers.py` and place it in a new module `peers_ch.py`.

We also update the `__init__.py` accordingly.

The internal search method `peers_ch.search_ch_peers`
is also used by `peers_multi.ch_search_ch_peers` (used by
`list_chs_peers` and `list_ch_subs_peers`)
so we adjust the import as well.