peers_claims: new method `list_peers` to search peers of a single claim
It needs to handle the cases when the claim doesn't exist,
or when the claim is not a stream (for example, a channel).

Only claims that are downloadable streams can have peers
in the network. Other types of claims will be processed
but won't display any `sd_hash`, and will have zero peers.

----

Formerly `list_peers` was used to search the peers
of multiple claims of a single channel, but this method
was renamed to `list_ch_peers` and moved to the `peers_ch.py` module.