peers_claims: new method `list_m_peers` to search peers of multiple claims
The input is a list of claims, by URI or claim ID,
it needs to handle the cases when the claims don't exist,
or when the claims are not a stream (for example, a channel).

Only claims that are downloadable streams can have peers
in the network. Other types of claims will be processed
but won't display any `sd_hash`, and will have zero peers.

----

This `list_m_peers` method functions similar to `peers_ch.list_ch_peers`,
the difference is that it uses a list of arbitrary claims
instead of the claims of a single channel.