funcs: use default value `fdate=False`
Many functions that use `funcs.print_content` set `fdate=None`,
however, this value should be a boolean value, so instead we set
it to `fdate=False`.

The default value should be
```
funcs.print_content(out, file=None, fdate=False)
```

The parameter `file=None` is correct because this one doesn't expect
a boolean but a string. We set it to `None` to indicate no string.