peers_base: new method `process_claims_peers` to process the peers
Given a dictionary, `base_peers_info`, which is the result
of the peer search of multiple claims, we can create a more complete
dictionary with more information.

The input `base_peers_info` variable includes only some keys
- `'n_claims'`
- `'n_streams'`
- `'streams_info'`

After the processing new keys are added
- `'total_size'`
- `'total_duration'`
- `'streams_with_hosts'`
- `'streams_with_hosts_all'`
- `'total_peers'`
- `'total_peers_all'`
- `'unique_nodes'`
- `'unique_trackers'`
- `'peer_ratio'`
- `'peer_ratio_all'`
- `'hosting_coverage'`
- `'hosting_coverage_all'`
- `'local_node'`

Moreover, if `channel=True` it will add another key `'channel'`
with the name of the channel, assuming it is included in `base_peers_info`.