peers_base: new method `get_summary` to collect information on peers
Given a dictionary with fully processed information, `peers_info`,
which is the result of the peer search of multiple claims,
we can create a summary text to be printed as a single paragraph.

This paragraph will contain information on the claims searched,
number of streams, number of user peers and tracker peers,
size and duration of all streams, and hosting ratio and hosting coverage.

With the parameter `channel=True`, we assume the `peer_info`
dictionary includes the `'channel'` key, so we can include it
in the summary. This is used when `peers_info` comes from
`peers_ch.search_ch_peers`, that is, from searching the peers
of the claims of a single channel.