peers_base: new method `get_claim_summary` to return a line summary
Given a dictionary with information, `stream_info`,
which is the result of the peer search of a single claim,
we can create a summary text to be printed as a single line, if `inline=True`,
or as a paragraph, if `inline=False`.

This paragraph will contain information on the claim searched,
number of user peers and tracker peers, size and duration of the claim,
and type of claim.

Only downloadable claims, that is, streams, will have peers, other types,
such as reposts or channels, won't have peers.

If `inline=True`, the default, a single line of information will be returned.
Thus this method can be used with a list of claims, to produce
a list of peer information, one for each claim.

If `inline=False`, then a full paragraph of information will be returned
for each claim.