peers_ch: use the new `process_claims_peers` in `search_ch_peers`
Instead of using a separate method, `process_ch_peers`,
we use `peers_base.process_claims_peers` with the `channel=True` option,
in order to process the peers found in each claim of a particular channel.

This method augments the base dictionary returned by `peers_base.calculate_peers`,
this augmented dictionary is the one that is used by `peers_base.get_m_summary`
and `print_p_lines` to print the final information on the peer search.